Introducing `before_upsert_with_payload` hook
This hook will allow applications to have access to both the resource
that is being materialized as well as the `payload`. This allows
applications to have custom loading logic for relationship elements. For
example, in `rider-api` we want to persist **deliveries** before we
persist `pickup` objects. This hook allows us to have access to the
`href` of deliveries and fetch them using the usual `materialist` flow,
but ensure that when application logic is execute we have all of the
relationships in place.